Refactor MetaMaskIdentity to support any wallet
