+++ packages/core/src/identities/index.js
@@ -1,2 +1,2 @@
 export { default as PrivateKeyIdentity } from './PrivateKeyIdentity';
-export { default as MetaMaskIdentity } from './MetaMaskIdentity';
+export { default as WalletRPCIdentity } from './WalletRPCIdentity';
